requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
benford-py==0.2.7
certifi==2020.6.20
click==7.1.2
colorama==0.4.3
cycler==0.10.0
Flask==1.1.2
flask-pytest==0.0.5
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
more-itertools==8.5.0
numpy==1.19.2
packaging==20.4
pandas==1.1.2
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.2
pytest-lazy-fixture==0.6.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
zipp==3.1.0
